Stripping down PlayerPRO 6 to make way for the rewrite. Prepare to work on PlayerPROKit. Use x86_64 exclusively on the apps and their plug-ins. This does not include PlayerPROCore and its plug-ins.
git-svn-id: svn://svn.code.sf.net/p/playerpro/code/trunk@951 f4605116-7794-4d0b-9e48-65ca9bfeedfd

+++ Files/(Plug-Ins)/Instruments Import:Export Plugs/WAV.c
@@ -84,7 +84,7 @@ static OSErr mainWave(void					*unused,
 			Ptr				theSound;
 			long			lS, lE;
 			short			sS;
-			unsigned long	rate;
+			unsigned int	rate;
 			Boolean			stereo;
 			size_t sndLen;
