+++ Files/(Plug-Ins)/Instruments Import:Export Plugs/WAV.h
@@ -107,7 +107,7 @@ __private_extern__ Ptr ConvertWAV(FSSpec *fileSpec, long *loopStart, long *loopE
 __private_extern__ OSErr ConvertDataToWAVE( FSSpec file, FSSpec *newfile, PPInfoPlug *thePPInfoPlug) DEPRECATED_ATTRIBUTE;
 #endif
 __private_extern__ OSErr TestWAV(PCMWavePtr CC);
-__private_extern__ Ptr ConvertWAVCFURL(CFURLRef theURL, size_t *sndSize, long *loopStart, long *loopEnd, short *sampleSize, unsigned long *rate, Boolean *stereo)
+__private_extern__ Ptr ConvertWAVCFURL(CFURLRef theURL, size_t *sndSize, long *loopStart, long *loopEnd, short *sampleSize, unsigned int *rate, Boolean *stereo)
 ;
 #ifdef __cplusplus
 }
